fix utils move
